[PROT-295] Extends infrastructure of entities for new feature implememntation

+++ src/main/java/io/protone/custom/service/mapper/CustomItemMapperExt.java
@@ -111,7 +111,7 @@ public class CustomItemMapperExt {
         }
         LibMediaItem db = new LibMediaItem()
             .idx(dto.getIdx())
-            .length(dto.getLength().longValue())
+            .length(dto.getLength().doubleValue())
             .name(dto.getName())
             .itemType(mapItemType(dto.getResourceType()))
             .state(LibItemStateEnum.valueOf(dto.getState().toString()))
